+++ backend/devices/bitbox02/device.go
@@ -28,16 +28,23 @@ import (
 	"github.com/sirupsen/logrus"
 )
 
-// ProductName is the name of the BitBox02 product.
-// If you change this, be sure to check the frontend and other places which assume this is a
+// If you change these, be sure to check the frontend and other places which assume they are
 // constant.
-const ProductName = "bitbox02"
+const (
+	// BitBox02ProductName is the name of the BitBox02 product.
+	BitBox02ProductName = "bitbox02"
+	// BitBox02NovaProductName is the name of the BitBox02 Nova product.
+	BitBox02NovaProductName = "bitbox02nova"
+	// PlatformName is the name of the BitVox02 platform.
+	PlatformName = "bitbox02"
+)
 
 // Device implements device.Device.
 type Device struct {
 	firmware.Device
-	deviceID string
-	log      *logrus.Entry
+	deviceID    string
+	productName string
+	log         *logrus.Entry
 
 	observable.Implementation
 }
@@ -51,10 +58,16 @@ func NewDevice(
 	communication firmware.Communication,
 	opts ...firmware.DeviceOption,
 ) *Device {
+	productName := BitBox02ProductName
+	// BitBox02Plus is the internal code name for the BitBox Nova.
+	if product == bitbox02common.ProductBitBox02PlusBTCOnly ||
+		product == bitbox02common.ProductBitBox02PlusMulti {
+		productName = BitBox02NovaProductName
+	}
 	log := logging.Get().
 		WithGroup("device").
 		WithField("deviceID", deviceID).
-		WithField("productName", ProductName).
+		WithField("productName", productName).
 		WithField("product", product)
 
 	log.Info("Plugged in device")
@@ -67,8 +80,9 @@ func NewDevice(
 			logger{log},
 			opts...,
 		),
-		deviceID: deviceID,
-		log:      log,
+		deviceID:    deviceID,
+		productName: productName,
+		log:         log,
 	}
 	device.Device.SetOnEvent(func(ev firmware.Event, meta interface{}) {
 		switch ev {
@@ -116,7 +130,12 @@ func (device *Device) init() {
 
 // ProductName implements device.Device.
 func (device *Device) ProductName() string {
-	return ProductName
+	return device.productName
+}
+
+// PlatformName implements device.Device.
+func (device *Device) PlatformName() string {
+	return PlatformName
 }
 
 // Identifier implements device.Device.