+++ frontends/web/src/components/dialog/dialog.tsx
@@ -19,7 +19,6 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { CloseXDark, CloseXWhite } from '@/components/icon';
 import { UseBackButton } from '@/hooks/backbutton';
 import { useEsc, useKeydown } from '@/hooks/keyboard';
-import { useMediaQuery } from '@/hooks/mediaquery';
 import style from './dialog.module.css';
 
 type TProps = {
@@ -51,7 +50,6 @@ export const Dialog = ({
   const modalRef = useRef<HTMLDivElement>(null);
   const modalContentRef = useRef<HTMLDivElement>(null);
   const timerIdRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const isMobile = useMediaQuery('(max-width: 768px)');
 
   const getFocusables = useCallback((): (NodeListOf<HTMLElement> | null) => {
     if (!modalContentRef.current) {
@@ -191,7 +189,7 @@ export const Dialog = ({
 
   const handleTap = (e: React.MouseEvent<HTMLDivElement>) => {
     const validTap = e.target === e.currentTarget;
-    if (validTap && isMobile) {
+    if (validTap) {
       closeHandler();
     }
   };